+++ ThirdPartyAdapters/ironsource/ironsource/src/main/java/com/google/ads/mediation/ironsource/IronSourceRewardedAvailabilityListener.java
@@ -1,0 +1,6 @@
+package com.google.ads.mediation.ironsource;
+
+public interface IronSourceRewardedAvailabilityListener {
+    void onRewardedAdAvailable();
+    void onRewardedAdNotAvailable();
+}