+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleManager.java
@@ -1,6 +1,5 @@
 package com.vungle.mediation;
 
-import android.content.Context;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
@@ -8,7 +7,6 @@ import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.vungle.warren.AdConfig;
-import com.vungle.warren.InitCallback;
 import com.vungle.warren.LoadAdCallback;
 import com.vungle.warren.PlayAdCallback;
 import com.vungle.warren.Vungle;
@@ -21,47 +19,32 @@ import java.util.Map;
  * A helper class to load and show Vungle ads and keep track of multiple
  * {@link VungleInterstitialAdapter} instances.
  */
-class VungleManager {
+public class VungleManager {
 
     private static final String TAG = VungleManager.class.getSimpleName();
     private static final String PLAYING_PLACEMENT = "placementID";
-    private static final String VERSION = "6.3.24";
 
     private static VungleManager sInstance;
     private String mCurrentPlayId = null;
-    private boolean mIsInitialising = false;
-    private String mAppId;
     private Handler mHandler = new Handler(Looper.getMainLooper());
     private Map<String, VungleListener> mListeners;
 
-    static VungleManager getInstance(String appId) {
+    public static VungleManager getInstance() {
         if (sInstance == null) {
-            sInstance = new VungleManager(appId);
+            sInstance = new VungleManager();
         }
         return sInstance;
     }
 
-    private VungleManager(String appId) {
+    private VungleManager() {
         mListeners = new HashMap<>();
 
         VungleApiClient.addWrapperInfo(VungleApiClient.WrapperFramework.admob,
-                VERSION.replace('.', '_'));
-
-        this.mAppId = appId;
-    }
-
-    boolean isInitialized() {
-        return Vungle.isInitialized();
-    }
-
-    @SuppressWarnings("SameParameterValue")
-    void setIncentivizedFields(String userID, String title, String body, String keepWatching,
-                               String close) {
-        Vungle.setIncentivizedFields(userID, title, body, keepWatching, close);
+                com.vungle.warren.BuildConfig.VERSION_NAME.replace('.', '_'));
     }
 
     @Nullable
-    String findPlacement(Bundle networkExtras, Bundle serverParameters) {
+    public String findPlacement(Bundle networkExtras, Bundle serverParameters) {
         String placement = null;
         if (networkExtras != null
                 && networkExtras.containsKey(VungleExtrasBuilder.EXTRA_PLAY_PLACEMENT)) {
@@ -74,76 +57,11 @@ class VungleManager {
             }
             placement = serverParameters.getString(PLAYING_PLACEMENT);
         }
-        if (placement == null) {
-            Log.e(TAG, "placementID not provided from serverParameters. Please check your AdMob dashboard settings." +
-                    "load and play functionality will not work");
-        }
         return placement;
     }
 
-    void init(Context context) {
-        if (Vungle.isInitialized()) {
-            for (VungleListener cb : mListeners.values()) {
-                if (cb.isWaitingInit()) {
-                    cb.setWaitingInit(false);
-                    cb.onInitialized(Vungle.isInitialized());
-                }
-            }
-            return;
-        }
-        if (mIsInitialising) {
-            return;
-        }
-        mIsInitialising = true;
-
-        Vungle.init(mAppId, context.getApplicationContext(), new InitCallback() {
-            @Override
-            public void onSuccess() {
-                mHandler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        mIsInitialising = false;
-                        if(VungleConsent.getCurrentVungleConsent() != null) {
-                            Vungle.updateConsentStatus(VungleConsent.getCurrentVungleConsent(),
-                                    VungleConsent.getCurrentVungleConsentMessageVersion());
-                        }
-                        for (VungleListener cb : mListeners.values()) {
-                            if (cb.isWaitingInit()) {
-                                cb.setWaitingInit(false);
-                                cb.onInitialized(Vungle.isInitialized());
-                            }
-                        }
-                    }
-                });
-            }
-
-            @Override
-            public void onError(Throwable throwable) {
-                mHandler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        mIsInitialising = false;
-                        for (VungleListener cb : mListeners.values()) {
-                            if (cb.isWaitingInit()) {
-                                cb.setWaitingInit(false);
-                                cb.onInitialized(Vungle.isInitialized());
-                            }
-                        }
-                    }
-                });
-            }
-
-            @Override
-            public void onAutoCacheAdAvailable(String placementId) {
-                // not used
-            }
-        });
-    }
-
     void removeListener(String id) {
-        if (mListeners.containsKey(id)) {
-            mListeners.remove(id);
-        }
+        mListeners.remove(id);
     }
 
     void addListener(String id, VungleListener listener) {
@@ -266,11 +184,6 @@ class VungleManager {
         }
     }
 
-    /**
-     * Checks and returns if the passed Placement ID is a valid placement for App ID
-     * @param placementId
-     * @return
-     */
     boolean isValidPlacement(String placementId) {
         return Vungle.isInitialized() &&
                 Vungle.getValidPlacements().contains(placementId);